// apps/tenant-dashboard/src/components/label/types.ts
import { BoxProps } from "@mui/material/Box";

// ----------------------------------------------------------------------

export type LabelColor =
  | "default"
  | "primary"
  | "secondary"
  | "info"
  | "success"
  | "warning"
  | "error";

export type LabelVariant = "filled" | "outlined" | "soft";

export interface LabelProps extends BoxProps {
  startIcon?: React.ReactElement | null;
  color?: LabelColor;
  variant?: LabelVariant;
}